Moved code around to enable easier testing
